% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{data_gen}
\alias{data_gen}
\title{data_gen}
\usage{

}
\arguments{
\item{n_doc}{The number of documents to generate}
}
\value{

}
\description{
second data generation function
}
\details{
TODO: finish later
}
