% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessments.R
\name{get_best_mapping}
\alias{get_best_mapping}
\title{get_best_mapping}
\usage{
get_best_mapping(predicted,true)
}
\arguments{
\item{predicted}{n length vector of predicted groups}

\item{true}{n length vector of true groups}
}
\value{
list("prop_correct","mapped_pred","true_val","mapping")
$prop_correct: float \link{0-1} proportion of the true value, that best prediction map correctly gets
$mapped_pred: n length vector, mapped predictions of best mapping
$true_val: n length vector of the true value
$mapping: k length vector (k is the number of groups) of the best group mappings
}
\description{
returns the best mapping through complete search
}
