% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessments.R
\name{get_plurarity_topics}
\alias{get_plurarity_topics}
\title{get_plurarity_topics}
\usage{
get_plurarity_topics(doc_2_top_mat)
}
\arguments{
\item{doc_2_top_mat}{n by k matrix of floating point numbers
The rows should sum to 1}
}
\value{
n-length vector of plurarity mappings
The values will be integers ranging from \link{1-K}
}
\description{
get the 1-indexed topics by plurarity of the document to topic matrix
}
