% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viLDAgibbs.R
\name{viLDA.gibbs}
\alias{viLDA.gibbs}
\title{viLDA.gibbs}
\usage{
viLDA.gibbs(N, K, p, data, alpha, beta, seed = 19890419, ...)
}
\arguments{
\item{N}{The number of documents}
}
\value{

}
\description{
viLDA.gibbs
}
