% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{data_gen}
\alias{data_gen}
\title{data_gen}
\usage{

}
\arguments{
\item{n_doc}{The number of documents to generate}
}
\value{

}
\description{
Technically this returns the number of "code points", in a string. One
code point usually corresponds to one character, but not always. For example,
an u with a umlaut might be represented as a single character or as the
combination a u and an umlaut.
}
