% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viLDA.R
\name{viLDA.stoch}
\alias{viLDA.stoch}
\title{viLDA.stoch}
\usage{
viLDA.stoch(
  dat,
  K,
  d = NULL,
  v = NULL,
  maxIterVal = NULL,
  maxVBiterVal = NULL,
  alphaWords = 0.2,
  alphaTopics = 0.2,
  rho = 0.001,
  tol = 0.1,
  seed = 19890419,
  ...
)
}
\arguments{
\item{dat}{The data in long format.
This is a data.frame with 3 columns:
$doc = doc ids
$words = word ids
$N = counts}
}
\value{

}
\description{
viLDA.stoch
}
